Remove the default derivative method from ActivationFunction
The derivative method should be implemented when the apply method is
implemented, because the derivative method depends on the apply method.
The risk of having a default implementation is that this will be
overlooked, while the benefit is low since it is not used on any of the
implementations of the ActivationFunction class at this time.