Introduce the MaxEvolutionStopCondition class
The current stop condition did not work well on the demo. A simpler stop
condition is needed. The MaxEvolutionStopCondition evaluates to false to
a given number of times and then evaluates to true. For this condition
we need to fetch the population from the stop condition, because the
population with the best chromosome may not be the last or the previous.