Alter mutation function to return chromosome
There are two problems. First, the fittest chromosomes are altered by
mutation causing the loss of the best solution at times. Second,
mutation causes previous mutations to mutate. These problems are
addressed by not altering the existing chromosome, but creating a new
chromosome instead and adding that to the new pool.